Microserviciu2/app/requirements.txt
fastapi
pymongo
uvicorn
pydantic
asyncpg
psycopg2-binary
pydantic[email]
httpx
grpcio
grpcio-tools
pyjwt